@@ -1,13 +1,18 @@
 {
   "name": "nino",
   "description": "Moderation bot for Discord's hack week entry.",
-  "version": "0.7.2",
+  "version": "0.8.0",
+  "homepage": "https://nino.augu.dev",
+  "license": "MIT",
+  "repository": "https://github.com/NinoDiscord/Nino",
+  "private": true,
+  "bugs": "https://github.com/NinoDiscord/Nino/issues/new?assignees=&labels=bug&template=bug_report.md&title=",
   "main": "dist/index.js",
   "scripts": {
-    "build:watch": "yarn lint && yarn clean && tsc -w",
-    "build": "yarn lint && yarn clean && tsc",
+    "build:watch": "yarn lint && yarn clean && yarn tsc -w",
+    "build": "yarn lint && yarn clean && yarn tsc",
     "clean": "rm -fr dist",
-    "main": "yarn build && node dist/index.js",
+    "main": "node dist/index.js",
     "lint": "eslint src --ext .ts --fix",
     "test": "jest --config jest.config.js --forceExit",
     "dev": "nodemon dist/index.js"
@@ -22,38 +27,35 @@
     "ohlookitsderpy"
   ],
   "dependencies": {
-    "@augu/immutable": "0.2.0",
-    "@sentry/node": "5.15.4",
+    "@augu/immutable": "0.2.5",
+    "@sentry/node": "5.18.0",
     "common-tags": "1.8.0",
-    "eris": "0.12.0",
+    "eris": "0.13.3",
     "inversify": "5.0.1",
     "inversify-inject-decorators": "3.1.0",
-    "ioredis": "4.16.2",
-    "js-yaml": "3.13.1",
+    "ioredis": "4.17.3",
+    "js-yaml": "3.14.0",
     "leeks.js": "0.0.8",
-    "mongoose": "5.9.9",
+    "mongoose": "5.9.20",
     "ms": "2.1.2",
-    "prom-client": "11.5.3",
     "reflect-metadata": "0.1.13",
     "wumpfetch": "0.3.1"
   },
   "devDependencies": {
-    "@augu/eslint-config": "1.4.0",
     "@types/common-tags": "1.8.0",
-    "@types/express": "4.17.6",
-    "@types/ioredis": "4.14.9",
-    "@types/jest": "25.2.1",
-    "@types/js-yaml": "3.12.3",
-    "@types/mongoose": "5.7.11",
+    "@types/ioredis": "4.16.7",
+    "@types/jest": "26.0.3",
+    "@types/js-yaml": "3.12.4",
+    "@types/mongoose": "5.7.27",
     "@types/ms": "0.7.31",
-    "@types/node": "13.11.1",
-    "@typescript-eslint/eslint-plugin": "2.28.0",
-    "@typescript-eslint/parser": "2.28.0",
-    "eslint": "6.8.0",
-    "jest": "25.3.0",
-    "jest-mock-extended": "1.0.8",
-    "nodemon": "2.0.3",
-    "ts-jest": "25.3.1",
-    "typescript": "3.8.3"
+    "@types/node": "13.13.5",
+    "@typescript-eslint/eslint-plugin": "3.4.0",
+    "@typescript-eslint/parser": "3.4.0",
+    "eslint": "7.3.1",
+    "jest": "26.1.0",
+    "jest-mock-extended": "1.0.9",
+    "nodemon": "2.0.4",
+    "ts-jest": "26.1.1",
+    "typescript": "3.9.5"
   }
 }
