@@ -12,20 +12,17 @@ export default class InviteCommand extends Command {
       name: 'invite',
       description: 'Gives you the invite for the bot.',
       aliases: ['inv'],
-      category: 'Generic',
       ownerOnly: false,
     });
   }
 
   async run(ctx: Context) {
     const embed = this.bot.getEmbed()
-      .setTitle('Invite Me')
-      .setDescription(stripIndents`
-        > **Invite me to your server today o-or join my owner's Discord to get updates and such...**
-
-        :link: **<https://discordapp.com/oauth2/authorize?client_id=${ctx.me.id}&scope=bot>**
-        :island: **<https://discord.gg/7TtMP2n>**
-      `)
+      .setTitle(ctx.translate('commands.generic.invite.title'))
+      .setDescription(ctx.translate('commands.generic.invite.description', {
+        invite: `<https://discordapp.com/oauth2/authorize?client_id=${ctx.me.id}&scope=bot>`,
+        server: 'https://discord.gg/7TtMP2n'
+      }))
       .build();
 
     return ctx.embed(embed);
