@@ -2,6 +2,7 @@ import { Punishment, PunishmentType } from '../../structures/managers/Punishment
 import { Constants, Member, Guild } from 'eris';
 import { injectable, inject } from 'inversify';
 import PermissionUtils from '../../util/PermissionUtils';
+import { Module } from '../../util';
 import { findId } from '../../util/UserUtil'; 
 import { TYPES } from '../../types';
 import Command from '../../structures/Command';
@@ -19,7 +20,7 @@ export default class BanCommand extends Command {
       description: 'Ban a member in the current guild',
       usage: '<user> <reason> [--soft] [--days]',
       aliases: ['banne', 'bean'],
-      category: 'Moderation',
+      category: Module.Moderation,
       guildOnly: true,
       userPermissions: Constants.Permissions.banMembers,
       botPermissions: Constants.Permissions.banMembers
@@ -27,16 +28,16 @@ export default class BanCommand extends Command {
   }
 
   async run(ctx: Context) {
-    if (!ctx.args.has(0)) return ctx.send('No user has been specified');
+    if (!ctx.args.has(0)) return ctx.sendTranslate('global.noUser');
 
     const userID = ctx.args.get(0);
     const user = findId(userID);
-    if (!user) return ctx.send('Unable to find that user. All users must be `<@mention>` or with a user ID');
+    if (!user) return ctx.sendTranslate('global.unableToFind');
 
     let member: Member | { id: string; guild: Guild } | undefined = ctx.guild!.members.get(user);
     if (!member || !(member instanceof Member)) member = { id: userID, guild: ctx.guild! };
     else {
-      if (!PermissionUtils.above(ctx.member!, member)) return ctx.send('User is above or the same of your heirarchy');
+      if (!PermissionUtils.above(ctx.member!, member)) return ctx.sendTranslate('global.heirarchy');
     }
 
     const baseReason = ctx.args.has(1) ? ctx.args.slice(1).join(' ') : undefined;
@@ -50,11 +51,11 @@ export default class BanCommand extends Command {
     }
 
     const days = ctx.flags.get('days') || ctx.flags.get('d');
-    if (days && (typeof days === 'boolean' || !(/[0-9]+/).test(days))) return ctx.send('You must need to specify an amount days to delete messages. (example: `--days=7`)');
+    if (days && (typeof days === 'boolean' || !(/[0-9]+/).test(days))) return ctx.sendTranslate('global.invalidFlag.string');
 
     const t = time ? ms(time) : undefined;
     const soft = ctx.flags.get('soft');
-    if (soft && typeof soft === 'string') return ctx.send('You don\'t need to append anything (example: `--soft`)');
+    if (soft && typeof soft === 'string') return ctx.sendTranslate('global.invalidFlag.boolean');
 
     const punishment = new Punishment(PunishmentType.Ban, {
       moderator: ctx.sender,
@@ -67,9 +68,18 @@ export default class BanCommand extends Command {
       await this.bot.punishments.punish(member!, punishment, reason);
 
       const prefix = member instanceof Member ? member.user.bot ? 'Bot' : 'User' : 'User';
-      return ctx.send(`${prefix} was successfully banned`);
+      return ctx.sendTranslate('commands.moderation.ban', {
+        type: prefix
+      });
     } catch(e) {
-      ctx.send(`Unable to ban user: \`${e.message}\``);
+      if (e.message.includes('snowflake')) return ctx.sendTranslate('commands.moderation.invalidSnowflake', {
+        type: 'banned'
+      });
+
+      return ctx.sendTranslate('commands.moderation.unable', {
+        type: member instanceof Member ? member.user.bot ? 'bot' : 'user' : 'user',
+        message: e.message
+      });
     }
   }
 }
