@@ -2,6 +2,7 @@ import { Punishment, PunishmentType } from '../../structures/managers/Punishment
 import { injectable, inject } from 'inversify';
 import { Constants, Member } from 'eris';
 import PermissionUtils from '../../util/PermissionUtils';
+import { Module } from '../../util';
 import { TYPES } from '../../types';
 import findUser from '../../util/UserUtil'; 
 import Command from '../../structures/Command';
@@ -17,8 +18,8 @@ export default class KickCommand extends Command {
       name: 'kick',
       description: 'Kicks a user from the guild',
       usage: '<user> <reason> [--reason]',
-      aliases: ['boot'],
-      category: 'Moderation',
+      aliases: ['boot', 'yeet'],
+      category: Module.Moderation,
       guildOnly: true,
       botPermissions: Constants.Permissions.kickMembers,
       userPermissions: Constants.Permissions.kickMembers
@@ -26,15 +27,18 @@ export default class KickCommand extends Command {
   }
 
   async run(ctx: Context) {
-    if (!ctx.args.has(0)) return ctx.send('No user has been specified');
+    if (!ctx.args.has(0)) return ctx.sendTranslate('global.noUser');
 
     const userID = ctx.args.get(0);
     const user = findUser(this.bot, userID);
-    if (!user || user === undefined) return ctx.send('Cannot find member in this guild');
+    if (!user || user === undefined) return ctx.sendTranslate('global.unableToFind');
 
     const member = ctx.guild!.members.get(user.id);
-    if (!member) return ctx.send(`User **\`${user.username}#${user.discriminator}\`** isn't in the guild`);
-    if (!PermissionUtils.above(ctx.message.member!, member)) return ctx.send('The user is above or the same heirarchy as you');
+    if (!member) return ctx.sendTranslate('commands.moderation.notInGuild', {
+      user: `${user.username}#${user.discriminator}`
+    });
+
+    if (!PermissionUtils.above(ctx.message.member!, member)) return ctx.sendTranslate('global.heirarchy');
 
     const reason = ctx.args.has(1) ? ctx.args.slice(1).join(' ') : undefined;
     const punishment = new Punishment(PunishmentType.Kick, {
@@ -45,9 +49,14 @@ export default class KickCommand extends Command {
       await this.bot.punishments.punish(member!, punishment, reason);
 
       const prefix = member instanceof Member ? member.user.bot ? 'Bot' : 'User' : 'User';
-      return ctx.send(`${prefix} was successfully kicked`);
+      return ctx.sendTranslate('commands.moderation.kick', {
+        type: prefix
+      });
     } catch(e) {
-      ctx.send(`Unable to kick user: \`${e.message}\``);
+      return ctx.sendTranslate('commands.moderation.unable', {
+        type: member instanceof Member ? member.user.bot ? 'bot' : 'user' : 'user',
+        message: e.message
+      });
     }
   }
 }
