@@ -1,5 +1,6 @@
 import { injectable, inject } from 'inversify';
 import { Constants } from 'eris';
+import { Module } from '../../util';
 import { TYPES } from '../../types';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
@@ -13,7 +14,7 @@ export default class ReasonCommand extends Command {
       description: 'Updates a case\'s reason',
       usage: '<caseID> <reason>',
       aliases: ['update'],
-      category: 'Moderation',
+      category: Module.Moderation,
       guildOnly: true,
       userPermissions: Constants.Permissions.banMembers,
       botPermissions: Constants.Permissions.manageMessages
@@ -21,8 +22,8 @@ export default class ReasonCommand extends Command {
   }
 
   async run(ctx: Context) {
-    if (!ctx.args.has(0)) return ctx.send('Missing the case ID to update');
-    if (!ctx.args.has(1)) return ctx.send('Missing the reason to update the case');
+    if (!ctx.args.has(0)) return ctx.sendTranslate('commands.moderation.reason.noCase');
+    if (!ctx.args.has(1)) return ctx.sendTranslate('commands.moderation.reason.noReason');
 
     const id = ctx.args.get(0);
     const reason = ctx.args.args.slice(1).join(' ');
@@ -30,7 +31,7 @@ export default class ReasonCommand extends Command {
     const _case = await this.bot.cases.get(ctx.guild!.id, parseInt(id));
     const settings = await this.bot.settings.get(ctx.guild!.id);
 
-    if (!_case || _case === null) return ctx.send(`Unable to find case **#${id}**, d-does it still exist?`);
+    if (!_case || _case === null) return ctx.sendTranslate('commands.moderation.reason.invalid', { id });
     _case.reason = reason;
 
     await this.bot.cases.update(ctx.guild!.id, parseInt(id), {
@@ -38,12 +39,15 @@ export default class ReasonCommand extends Command {
         'reason': reason
       }
     }, async (error) => {
-      if (error) return ctx.send(`Unable to update case **#${id}**`);
+      if (error) return ctx.sendTranslate('commands.moderation.reason.error', { id });
 
       const message = await this.bot.client.getMessage(settings!.modlog, _case.message);
       await this.bot.punishments.editModlog(_case, message);
 
-      return ctx.send(`:ok_hand: Updated case **#${id}** for **${reason}**`);
+      return ctx.sendTranslate('commands.moderation.reason.edited', {
+        reason,
+        id
+      });
     });
   }
 }
