@@ -1,14 +1,12 @@
-import {
-  Punishment,
-  PunishmentType,
-} from '../../structures/managers/PunishmentManager';
+import { Punishment, PunishmentType } from '../../structures/managers/PunishmentManager';
+import { injectable, inject } from 'inversify';
 import { Constants } from 'eris';
-import Bot from '../../structures/Bot';
 import { findId } from '../../util/UserUtil';
+import { Module } from '../../util';
+import { TYPES } from '../../types';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
-import { injectable, inject } from 'inversify';
-import { TYPES } from '../../types';
+import Bot from '../../structures/Bot';
 
 @injectable()
 export default class UnbanCommand extends Command {
@@ -18,7 +16,7 @@ export default class UnbanCommand extends Command {
       description: 'Unbans a user from a guild',
       usage: '<user> <reason> [--reason]',
       aliases: ['unbanne'],
-      category: 'Moderation',
+      category: Module.Moderation,
       guildOnly: true,
       userPermissions: Constants.Permissions.banMembers,
       botPermissions: Constants.Permissions.banMembers
@@ -26,28 +24,28 @@ export default class UnbanCommand extends Command {
   }
 
   async run(ctx: Context) {
-    if (!ctx.args.has(0))
-      return ctx.send('Sorry but you will need to specify a user.');
+    if (!ctx.args.has(0)) return ctx.sendTranslate('global.noUser');
 
     const id = findId(ctx.args.get(0));
 
-    if (!id)
-      return ctx.send('Please type the id or mention the user (<@id>/<@!id>)');
-
-    if (!(await ctx.guild!.getBans()).find(v => v.user.id === id))
-      return ctx.send('The user is not banned from this guild.');
+    if (!id) return ctx.sendTranslate('global.unableToFind');
+    if (!(await ctx.guild!.getBans()).find(v => v.user.id === id)) return ctx.sendTranslate('global.notInBanList');
 
     const reason = ctx.args.has(1) ? ctx.args.slice(1).join(' ') : undefined;
     await this.bot.timeouts.cancelTimeout(id, ctx.guild!, 'unban');
 
     const punishment = new Punishment(PunishmentType.Unban, {
-      moderator: ctx.sender,
+      moderator: ctx.sender
     });
+    
     try {
       await this.bot.punishments.punish({ id, guild: ctx.guild! }, punishment, reason);
-      return ctx.send('User was successfully unbanned');
+      return ctx.sendTranslate('commands.moderation.unban');
     } catch (e) {
-      ctx.send('Cannot unban user, ' + e.message);
+      return ctx.sendTranslate('commands.moderation.unable', {
+        type: 'User',
+        message: e.message
+      });
     }
   }
 }
