@@ -1,7 +1,8 @@
 import { injectable, inject } from 'inversify';
 import { Constants } from 'eris';
-import findUser from '../../util/UserUtil';
+import { Module } from '../../util';
 import { TYPES } from '../../types';
+import findUser from '../../util/UserUtil';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
 import Bot from '../../structures/Bot';
@@ -14,10 +15,9 @@ export default class WarningsCommand extends Command {
       description: 'Shows the amount of warnings a member has.',
       usage: '<user>',
       aliases: ['warns'],
-      category: 'Moderation',
-      userPermissions:
-        Constants.Permissions.manageRoles |
-        Constants.Permissions.manageChannels,
+      category: Module.Moderation,
+      userPermissions: Constants.Permissions.manageRoles,
+      botPermissions: Constants.Permissions.manageRoles,
       guildOnly: true
     });
   }
@@ -25,14 +25,19 @@ export default class WarningsCommand extends Command {
   async run(ctx: Context) {
     if (!ctx.args.has(0)) return ctx.send('You need to specify a user.');
 
-    const u = findUser(this.bot, ctx.args.get(0))!;
-    if (!u) return ctx.send('I can\'t find this user!');
-    const member = ctx.guild!.members.get(u.id);
+    const userID = ctx.args.get(0);
+    const u = findUser(this.bot, userID);
+    if (!u || u === undefined) return ctx.sendTranslate('global.unableToFind');
 
-    if (!member) return ctx.send(`User \`${u.username}#${u.discriminator}\` is not in this guild?`);
+    const member = ctx.guild!.members.get(u.id);
+    if (!member) return ctx.sendTranslate('commands.moderation.notInGuild', {
+      user: `${u.username}#${u.discriminator}`
+    });
 
     const settings = await this.bot.warnings.get(ctx.guild!.id, member.id);
-    if (!settings) return ctx.send(`${member.username}#${member.discriminator} doesn't have any warnings in **${ctx.guild!.name}**`);
-    else return ctx.send(`${member.username}#${member.discriminator} has ${settings.amount} warnings to their name in **${ctx.guild!.name}**`);
+
+    return !settings
+      ? ctx.sendTranslate('commands.moderation.warnings.none')
+      : ctx.sendTranslate('commands.moderation.warnings.message', { warnings: settings.amount });
   }
 }
