@@ -15,7 +15,6 @@ export default class GuildJoinedEvent extends Event {
     this.bot.logger.info(`Joined guild ${guild.name} (${guild.id})`);
 
     this.bot.status.updateStatus();
-    this.bot.prometheus.guildCount.inc();
     this.bot.statistics.guildCount++;
     await this.bot.redis.set('guilds', this.bot.client.guilds.size);
 
@@ -24,7 +23,7 @@ export default class GuildJoinedEvent extends Event {
       const chan = (channel as TextChannel);
       const embed = this.bot.getEmbed()
         .setAuthor(`| Joined ${guild.name} (${guild.id})`, undefined, this.bot.client.user.dynamicAvatarURL('png', 1024))
-        .setFooter(`Now at ${this.bot.client.guilds.size} Guilds`, this.bot.client.user.dynamicAvatarURL('png', 1024))
+        .setFooter(`Now at ${this.bot.client.guilds.size} Guilds`)
         .build();
 
       chan.createMessage({ embed });
