@@ -1,5 +1,5 @@
 import { injectable, inject } from 'inversify';
-import { Message } from 'eris';
+import { Message, TextChannel } from 'eris';
 import { TYPES } from '../types';
 import Client from '../structures/Bot';
 import Event from '../structures/Event';
@@ -12,7 +12,7 @@ export default class MessageReceivedEvent extends Event {
     super(client, 'messageCreate');
   }
 
-  async emit(m: Message) {
+  async emit(m: Message<TextChannel>) {
     this.bot.manager.service.handle(m);
     this.bot.automod.handleMessage(m);
   }
