@@ -18,11 +18,9 @@ export default class ReadyEvent extends Event {
     this.bot.status.updateStatus();
     setInterval(() => this.bot.status.updateStatus(), 600000);
 
-    this.bot.prometheus.server.listen(this.bot.config.prometheus, () => this.bot.logger.info(`Metrics server is now listening at localhost:${this.bot.config.prometheus}`));
+    if (this.bot.client.user.id === '531613242473054229') this.bot.botlists.start();
     await this.bot.redis.set('guilds', this.bot.client.guilds.size);
-
     this.bot.statistics.guildCount = this.bot.client.guilds.size;
-    this.bot.prometheus.guildCount.set(this.bot.statistics.guildCount, Date.now());
     this.bot.timeouts.reapplyTimeouts();
   }
 }
