@@ -1,4 +1,6 @@
 import { injectable, unmanaged } from 'inversify';
+import { stripIndents } from 'common-tags';
+import { Module } from '../util';
 import Context from './Context';
 import Client from './Bot';
 import 'reflect-metadata';
@@ -7,7 +9,7 @@ export interface CommandInfo {
   name: string;
   description: string | ((client: Client) => string);
   usage?: string;
-  category?: string;
+  category?: Module;
   aliases?: string[];
   guildOnly?: boolean;
   ownerOnly?: boolean;
@@ -24,7 +26,7 @@ export default abstract class NinoCommand {
   public name: string;
   public description: string;
   public usage: string;
-  public category: string;
+  public category: Module;
   public aliases: string[];
   public guildOnly: boolean;
   public ownerOnly: boolean;
@@ -34,10 +36,7 @@ export default abstract class NinoCommand {
   public botPermissions: number;
   public userPermissions: number;
 
-  constructor(
-    client: Client, 
-    @unmanaged() info: CommandInfo
-  ) {
+  constructor(client: Client, @unmanaged() info: CommandInfo) {
     this.bot = client;
     this.name = info.name;
     this.description =
@@ -45,7 +44,7 @@ export default abstract class NinoCommand {
         ? info.description(client)
         : info.description;
     this.usage = info.usage || '';
-    this.category = info.category || 'Generic';
+    this.category = info.category || Module.Generic;
     this.aliases = info.aliases || [];
     this.guildOnly = info.guildOnly || false;
     this.ownerOnly = info.ownerOnly || false;
@@ -56,23 +55,41 @@ export default abstract class NinoCommand {
     this.userPermissions = info.userPermissions || 0;
   }
 
-  public abstract run(ctx: Context): Promise<any>;
+  public abstract run(ctx: Context, ...args: any[]): Promise<any>;
 
   format() {
     return `${this.bot.config.discord.prefix}${this.name}${this.usage ? ` ${this.usage}` : ''}`;
   }
 
-  help() {
+  async help(ctx: Context) {
+    const getPart = (type:
+      'title' |
+      'name' |
+      'syntax' |
+      'category' |
+      'aliases' |
+      'guildOnly' |
+      'ownerOnly' |
+      'cooldown'
+    ) => ctx.translate(`commands.generic.help.command.${type}`);
+
+    const getGlobalPart = (type: 'yes' | 'no') => ctx.translate(`global.${type}`);
+
     const embed = this.bot
       .getEmbed()
-      .setTitle(`Command ${this.name}`)
-      .setDescription(`**${this.description}**`)
-      .addField('Syntax', this.format(), true)
-      .addField('Category', this.category, true)
-      .addField('Aliases', this.aliases.length > 1 ? this.aliases.join(', ') : 'No aliases available', true)
-      .addField('Guild Only', this.guildOnly ? 'Yes' : 'No', true)
-      .addField('Owner Only', this.ownerOnly ? 'Yes' : 'No', true)
-      .addField('Cooldown', `${this.cooldown} seconds`, true);
+      .setTitle(ctx.translate('commands.generic.help.command.title', { command: this.name }))
+      .setDescription(stripIndents`
+        **${this.description}**
+
+        \`\`\`apache
+        ${getPart('name')}:       ${this.name}
+        ${getPart('syntax')}:     ${this.format()}
+        ${this.aliases.length ? `${getPart('aliases')}:    ${this.aliases.join(', ')}` : `${getPart('aliases')}:    None`}
+        ${getPart('guildOnly')}: ${this.guildOnly ? getGlobalPart('yes') : getGlobalPart('no')}
+        ${getPart('ownerOnly')}: ${this.ownerOnly ? getGlobalPart('yes') : getGlobalPart('no')}
+        ${getPart('cooldown')}:   ${this.cooldown} second${this.cooldown > 1 ? 's' : ''}
+        \`\`\`
+      `);
 
     return embed.build();
   }
