@@ -17,6 +17,7 @@ export interface Embed {
 export interface EmbedAuthor {
   name: string;
   url?: string;
+  // eslint-disable-next-line camelcase
   icon_url?: string;
 }
 
@@ -36,6 +37,7 @@ export interface EmbedField {
 
 export interface EmbedFooter {
   text: string;
+  // eslint-disable-next-line camelcase
   icon_url?: string;
 }
 
@@ -95,6 +97,7 @@ export default class EmbedBuilder {
     this.author = {
       name,
       url,
+      // eslint-disable-next-line camelcase
       icon_url: iconURL,
     };
 
@@ -107,12 +110,9 @@ export default class EmbedBuilder {
   }
 
   addField(name: string, value: string, inline: boolean = false) {
-    if (!name)
-      throw new Error('You d-didn\'t set a name to the field you baka!!');
-    if (!value)
-      throw new Error('You d-didn\'t set a value to the field you baka!!');
-    if (this.fields!.length > 25)
-      throw new Error('Unable to add anymore fields. (FIELD_LIMIT_THRESHOLD)');
+    if (!name) throw new Error('You d-didn\'t set a name to the field you baka!!');
+    if (!value) throw new Error('You d-didn\'t set a value to the field you baka!!');
+    if (this.fields!.length > 25) throw new Error('Unable to add anymore fields. (FIELD_LIMIT_THRESHOLD)');
 
     this.fields!.push({ name, value: toString(value), inline });
     return this;
@@ -129,6 +129,7 @@ export default class EmbedBuilder {
   }
 
   setFooter(txt: string, iconURL?: string) {
+    // eslint-disable-next-line camelcase
     this.footer = { text: txt, icon_url: iconURL };
     return this;
   }
@@ -147,6 +148,7 @@ export default class EmbedBuilder {
         ? {
           name: this.author.name,
           url: this.author.url,
+          // eslint-disable-next-line camelcase
           icon_url: this.author.icon_url
         }
         : undefined,
@@ -154,6 +156,7 @@ export default class EmbedBuilder {
       footer: this.footer
         ? {
           text: this.footer.text,
+          // eslint-disable-next-line camelcase
           icon_url: this.footer.icon_url
         }
         : undefined,
