@@ -1,5 +1,4 @@
 import { Message, TextChannel } from 'eris';
-import { replaceMessage } from '../../util';
 import PermissionUtils from '../../util/PermissionUtils';
 import Bot from '../Bot';
 
@@ -23,34 +22,38 @@ export default class AutoModBadWords {
    *
    * @param m the message
    */
-  async handle(m: Message): Promise<boolean> {
-    const channel = m.channel as TextChannel;
-    const guild = channel.guild;
-    const me = guild.members.get(this.bot.client.user.id)!;
+  async handle(m: Message<TextChannel>) {
+    if (!m || m === null) return false;
+    if (!(m.channel instanceof TextChannel)) return false;
+
+    const me = m.channel.guild.members.get(this.bot.client.user.id)!;
 
     if (
       !PermissionUtils.above(me, m.member!) ||
-      !channel.permissionsOf(me.id).has('manageMessages') ||
+      !m.channel.permissionsOf(me.id).has('manageMessages') ||
       m.author.bot ||
-      channel.permissionsOf(m.author.id).has('manageMessages')
+      m.channel.permissionsOf(m.author.id).has('manageMessages')
     ) return false;
 
-    const settings = await this.bot.settings.get(guild.id);
+    const settings = await this.bot.settings.get(m.channel.guild.id);
     if (
       !settings ||
       !settings.automod.badwords.enabled
     ) return false;
 
+    const user = await this.bot.userSettings.get(m.author.id);
+    const locale = user === null
+      ? this.bot.locales.get(settings.locale)!
+      : this.bot.locales.get(user.locale)!;
+
     for (let word of settings.automod.badwords.wordlist) {
       if (m.content.toLowerCase().indexOf(word.toLowerCase()) !== -1) {
         const punishments = await this.bot.punishments.addWarning(m.member!);
-        const response = !settings.responses.badwords.enabled ?
-          replaceMessage('Watch your language, %author%', m.author) :
-          replaceMessage(settings.responses.badwords.message, m.author);
+        const response = locale.translate('automod.badwords', { user: m.member ? `${m.member.username}#${m.member.discriminator}` : `${m.author.username}#${m.author.discriminator}` });
 
         await m.channel.createMessage(response);
         await m.delete();
-        for (let punishment of punishments) await this.bot.punishments.punish(m.member!, punishment, 'Automod (Swearing)');
+        for (let punishment of punishments) await this.bot.punishments.punish(m.member!, punishment, '[Automod] Swearing');
         return true;
       }
     }
