@@ -1,7 +1,6 @@
 import { Message, TextChannel } from 'eris';
-import { replaceMessage } from '../../util';
 import PermissionUtils from '../../util/PermissionUtils';
-import RedisQueue from '../../util/RedisQueue';
+import { Collection } from '@augu/immutable';
 import Bot from '../Bot';
 
 /**
@@ -15,9 +14,38 @@ import Bot from '../Bot';
  */
 export default class AutoModSpam {
   public bot: Bot;
+  public buckets: Collection<Collection<number[]>> = new Collection<Collection<number[]>>();
 
   constructor(client: Bot) {
     this.bot = client;
+  }
+
+  __cleanUp(guildId: string) {
+    let now = Date.now();
+    let ids: (string | number)[] = [];
+    this.buckets.get(guildId)!.forEach((v, k) => {
+      let diff = now - v[v.length-1];
+      if (now - v[v.length - 1] >= 5000) {
+        ids.push(k);
+      }
+    });
+    ids.forEach(element => {
+      this.buckets.delete(element);
+    });
+  }
+
+  __getQueue(guildId: string, userId: string): number[] {
+    if (!this.buckets.has(guildId)) {
+      this.buckets.set(guildId, new Collection<number[]>());
+    }
+    if (!this.buckets.get(guildId)!.has(userId)) {
+      this.buckets.get(guildId)!.set(userId, []);
+    }
+    return this.buckets.get(guildId)!.get(userId)!;
+  }
+
+  __clearQueue(guildId: string, userId: string) {
+    this.buckets.get(guildId)!.delete(userId);
   }
 
   /**
@@ -29,28 +57,30 @@ export default class AutoModSpam {
    *
    * @param m the message
    */
-  async handle(m: Message): Promise<boolean> {
-    const guild = (m.channel as TextChannel).guild;
-    const me = guild.members.get(this.bot.client.user.id)!;
+  async handle(m: Message<TextChannel>): Promise<boolean> {
+    if (!m || m === null) return false;
+    if (!(m.channel instanceof TextChannel)) return false;
+
+    const me = m.channel.guild.members.get(this.bot.client.user.id)!;
+
     if (
       !PermissionUtils.above(me, m.member!) ||
+      !m.channel.permissionsOf(me.id).has('manageMessages') ||
       m.author.bot ||
-      (m.channel as TextChannel)
-        .permissionsOf(m.author.id)
-        .has('manageMessages')
+      m.channel.permissionsOf(m.author.id).has('manageMessages')
     ) return false;
 
-    const settings = await this.bot.settings.get(guild.id);
+    const settings = await this.bot.settings.get(m.channel.guild.id);
     if (!settings || !settings.automod.spam) return false;
 
-    const queue = new RedisQueue(this.bot.redis, `${m.author.id}:${guild.id}`);
-    await queue.push(m.timestamp.toString());
+    const queue = this.__getQueue(m.channel.guild.id, m.author.id);
+    queue.push(m.timestamp);
 
-    const length = await queue.length();
-    if (length >= 5) {
-      const oldTime = Number.parseInt(await queue.pop());
+    if (queue.length >= 5) {
+      const oldTime = queue.shift()!;
       if (m.editedTimestamp && m.editedTimestamp > m.timestamp) return false;
       if (m.timestamp - oldTime <= 3000) {
+        this.__clearQueue(m.channel.guild.id, m.author.id);
         let punishments = await this.bot.punishments.addWarning(m.member!);
         for (let punishment of punishments) await this.bot.punishments.punish(
           m.member!,
@@ -58,14 +88,18 @@ export default class AutoModSpam {
           `[Automod] Spamming in <#${m.channel.id}>`
         );
 
-        const response = !settings.responses.spam.enabled ?
-          replaceMessage('Spamming is not allowed, %author%', m.author) :
-          replaceMessage(settings.responses.spam.message, m.author);
-
+        const user = await this.bot.userSettings.get(m.author.id);
+        const locale = user === null
+          ? this.bot.locales.get(settings.locale)!
+          : this.bot.locales.get(user.locale)!;
+  
+        const response = locale.translate('automod.invites', { user: m.member ? `${m.member.username}#${m.member.discriminator}` : `${m.author.username}#${m.author.discriminator}` });  
         await m.channel.createMessage(response);
         return true;
       }
     }
+    
+    this.__cleanUp(m.channel.guild.id);
 
     return false;
   }
