@@ -6,35 +6,9 @@ import mongoose from 'mongoose';
 import Logger from '../Logger';
 import 'reflect-metadata';
 
-interface BuildInfo {
-  version: string;
-  gitVersion: string;
-  modules: any[];
-  allocator: string;
-  javascriptEngine: string;
-  sysInfo: 'deprecated';
-  versionArray: number[];
-  debug: false;
-  maxBsonObjectSize: number;
-  storageEngines: string[];
-  ok: number;
-  openssl: { running: string; compiled: string };
-  buildEnvironment: {
-    distmod: string;
-    distarch: string;
-    cc: string;
-    ccflags: string;
-    cxx: string;
-    linkflags: string;
-    target_arch: string;
-    target_os: string;
-  };
-}
-
 @injectable()
 export default class DatabaseManager {
   public admin!: Admin;
-  public build!: BuildInfo;
   public logger: Logger = new Logger();
   public uri: string;
   public m!: typeof mongoose;
@@ -54,16 +28,12 @@ export default class DatabaseManager {
 
     this.m.connection.on('error', error => error ? this.logger.error(error) : null);
     this.logger.database(`Opened a connection to MongoDB with URI: ${this.uri}`);
+
+    this.admin = this.m.connection.db.admin();
   }
 
   dispose() {
     this.m.connection.close();
     this.logger.warn('Database connection was disposed');
-  }
-
-  async getBuild() {
-    if (!this.admin) this.admin = this.m.connection.db.admin();
-    if (!this.build) this.build = await this.admin.buildInfo();
-    return this.build;
   }
 }
