@@ -3,8 +3,10 @@ import Bot, { Config } from '../Bot';
 import { TYPES } from '../../types';
 import { Message, TextableChannel, User } from 'eris';
 import CommandContext from '../Context';
+import Language from '../Language';
 
 describe('CommandService', () => {
+
   container.rebind<Config>(TYPES.Config).toConstantValue({
     status: undefined,
     statusType: undefined,
@@ -38,7 +40,7 @@ describe('CommandService', () => {
   it('it should create a command invocation', () => {
     const message = { content: '!help hi' } as Message;
     const args = ['help', 'hi'];
-    const context = new CommandContext(bot, message, args);
+    const context = new CommandContext(bot, message, args, bot.locales.get('en_US'), undefined);
     expect(context).toBeDefined();
     expect(context.args).toBeDefined();
     expect(context.args.args).toEqual(expect.arrayContaining(args));
@@ -50,7 +52,7 @@ describe('CommandService', () => {
   it('it should not create a command invocation', () => {
     const message = { content: '' } as Message;
     const args = [];
-    const context = new CommandContext(bot, message, args);
+    const context = new CommandContext(bot, message, args, bot.locales.get('en_US'), undefined);
     expect(context).toBeDefined();
     expect(context.args).toBeDefined();
     expect(context.args.args).toEqual(expect.arrayContaining(args));
@@ -61,7 +63,7 @@ describe('CommandService', () => {
   it('it should be able to invoke the command', () => {
     const message = { content: '!help hi' } as Message;
     const args = ['help', 'hi'];
-    const context = new CommandContext(bot, message, args);
+    const context = new CommandContext(bot, message, args, bot.locales.get('en_US'), undefined);
     const invocation = commandService.getCommandInvocation(context);
     expect(invocation).toBeDefined();
     expect(invocation!.canInvoke()).toBeUndefined();
@@ -71,10 +73,12 @@ describe('CommandService', () => {
     bot.manager.getCommand('help')!.disabled = true;
     const message = { content: '!help hi' } as Message;
     const args = ['help', 'hi'];
-    const context = new CommandContext(bot, message, args);
+    const context = new CommandContext(bot, message, args, bot.locales.get('en_US'), undefined);
     const invocation = commandService.getCommandInvocation(context);
     expect(invocation).toBeDefined();
-    expect(invocation!.canInvoke()).toEqual('Currently, command `help` is globally disabled');
+    expect(invocation!.canInvoke()).toBeDefined();
+    expect(invocation!.canInvoke()!.key).toEqual('errors.disabled');
+    expect(invocation!.canInvoke()!.args).toEqual({ command: 'help' });
   });
 
   it('it should not be able to invoke the command because the command is guild only', () => {
@@ -83,12 +87,12 @@ describe('CommandService', () => {
       channel: { type: 1 } as TextableChannel,
     } as Message;
     const args = ['settings'];
-    const context = new CommandContext(bot, message, args);
+    const context = new CommandContext(bot, message, args, bot.locales.get('en_US'), undefined);
     const invocation = commandService.getCommandInvocation(context);
     expect(invocation).toBeDefined();
-    expect(invocation!.canInvoke()).toEqual(
-      'Sorry, but you need to be in a guild to execute the `settings` command.'
-    );
+    expect(invocation!.canInvoke()).toBeDefined();
+    expect(invocation!.canInvoke()!.key).toEqual('errors.guildOnly');
+    expect(invocation!.canInvoke()!.args).toEqual({ command: 'settings' });
   });
 
   it('it should not be able to invoke the command because the command is owner only', () => {
@@ -97,11 +101,11 @@ describe('CommandService', () => {
       author: { id: '1' } as User,
     } as Message;
     const args = ['eval'];
-    const context = new CommandContext(bot, message, args);
+    const context = new CommandContext(bot, message, args, bot.locales.get('en_US'), undefined);
     const invocation = commandService.getCommandInvocation(context);
     expect(invocation).toBeDefined();
-    expect(invocation!.canInvoke()).toEqual(
-      'Sorry, but you need to be a developer to execute the `eval` command.'
-    );
+    expect(invocation!.canInvoke()).toBeDefined();
+    expect(invocation!.canInvoke()!.key).toEqual('errors.ownerOnly');
+    expect(invocation!.canInvoke()!.args).toEqual({ command: 'eval' });
   });
 });
