@@ -14,6 +14,7 @@ export const TYPES = {
   GuildSettings: Symbol.for('GuildSettings'),
   BotListService: Symbol.for('BotListService'),
   StatusManager: Symbol.for('StatusManager'),
-  PrometheusManager: Symbol.for('PrometheusManager'),
-  CommandStatisticsManager: Symbol.for('CommandStatisticsManager')
+  CommandStatisticsManager: Symbol.for('CommandStatisticsManager'),
+  UserSettings: Symbol.for('UserSettings'),
+  GuildCounterManager: Symbol.for('GuildCounterManager')
 };
